Custom Shell Commands
You can use it with the endpoint in the api, and adding those commands inside the customCommand object in the config of Remote Control.  This still  an insecure way of implement this feature, so in the next update the remote endpoint will be deprecated. Try to use the API endpoints instead. Cheers! #159